Repaired out of bounds bug in Word.evaluate_word() by replacing the for loop with a while
It seems that you cannot "permanently" increment/decrement the index variable in python during execution, since the range(start, begin) creates a list of integers that the loop will use as the value for index. Therefore, index = index - 1 only decremented the variable until the end of the cycle. When the cicle ended, the variable would take the next value in the list, basically becoming index = index + 2
For example: index = 4 at the beggining of the cycle, index = index - 1 will make index take the value 3 for the rest of the cycle, but at the start of the next cycle in the loop, the value of index will be 5